requirements.txt
flask
flask-bcrypt
flask-jwt-extended
psycopg2-binary
gunicorn
flask_cors